+++ test/integration/200-get-delete.js
@@ -1,0 +1,31 @@
+import { strict as assert } from "node:assert";
+import { after, describe, it } from "node:test";
+
+import { mockAgent } from "@indiekit-test/mock-agent";
+import { testServer } from "@indiekit-test/server";
+import { testCookie } from "@indiekit-test/session";
+import { JSDOM } from "jsdom";
+import supertest from "supertest";
+
+await mockAgent("endpoint-posts");
+const server = await testServer({
+  application: { micropubEndpoint: "https://micropub-endpoint.example" },
+});
+const request = supertest.agent(server);
+
+describe("endpoint-posts GET /posts/:uid/delete", () => {
+  it("Gets delete confirmation page", async () => {
+    const response = await request
+      .get(`/posts/123/delete`)
+      .set("cookie", testCookie());
+    const dom = new JSDOM(response.text);
+    const result = dom.window.document.querySelector("title").textContent;
+
+    assert.equal(
+      result,
+      "Are you sure you want to delete this post? - Test configuration",
+    );
+  });
+
+  after(() => server.close());
+});